财产模型+修改实体ID由int 变为 long 还未实际测过，但allTests ok

+++ src/main/java/manager/dao/UserDAO.java
@@ -6,43 +6,42 @@ import manager.entity.general.User;
 import manager.entity.general.UserGroup;
 import manager.exception.DBException;
 import manager.exception.NoSuchElement;
-import manager.system.SMDB;
 import manager.system.SMError;
 import manager.system.SMPerm;
 
 public interface UserDAO {
 	
-	int insertUser(User user) throws DBException;
+	long insertUser(User user) throws DBException;
 	void updateExistedUser(User user) throws DBException;
-	User selectUser(int id) throws NoSuchElement, DBException;
+	User selectUser(long id) throws NoSuchElement, DBException;
 	User selectUniqueUserByField(String field,Object val) throws DBException,NoSuchElement;
 	boolean includeUniqueUserByField(String fieldName,String fieldVal) throws DBException;
 	
-	boolean includeUserGroup(int groupId) throws DBException;
+	boolean includeUserGroup(long groupId) throws DBException;
 	boolean includeUniqueUserGroupByField(String field,String val) throws DBException;
-	List<Integer> selectGroupsByUser(int userId) throws DBException;;
-	List<Integer> selectPermsByGroup(int groupId) throws DBException;;
-	List<Integer> selectUsersIdByGroup(int groupId) throws DBException;;
-	List<User> selectUsersByGroup(int groupId,int limit) throws DBException;
+	List<Long> selectGroupsByUser(long userId) throws DBException;;
+	List<Integer> selectPermsByGroup(long groupId) throws DBException;;
+	List<Long> selectUsersIdByGroup(long groupId) throws DBException;;
+	List<User> selectUsersByGroup(long groupId,long limit) throws DBException;
 	
 	List<UserGroup> selectAllUserGroup() throws DBException;
 	
-	long countUsersOfGroup(int groupId) throws DBException;
+	long countUsersOfGroup(long groupId) throws DBException;
 	long countAllUsers() throws DBException;
 	
-	int insertUserGroup(UserGroup group) throws DBException;
+	long insertUserGroup(UserGroup group) throws DBException;
 	UserGroup selectUniqueExistedUserGroupByField(String field,String val)throws DBException;
 	void updateExistedUserGroup(UserGroup group) throws DBException;
 	
-	void insertUsersToGroup(List<Integer> usersId, int groupId) throws DBException;
-	void deleteUsersFromGroup(List<Integer> usersId, int groupId) throws DBException;
+	void insertUsersToGroup(List<Long> usersId, long groupId) throws DBException;
+	void deleteUsersFromGroup(List<Long> usersId, long groupId) throws DBException;
 	
-	void insertPermsToGroup(List<SMPerm> perms,int groupId) throws DBException;
+	void insertPermsToGroup(List<SMPerm> perms,long groupId) throws DBException;
 	
-	void deletePermsFromGroup(List<SMPerm> perms,int groupId) throws DBException;
+	void deletePermsFromGroup(List<SMPerm> perms,long groupId) throws DBException;
 	
 	
-	default User selectExistedUser(int id) throws DBException {
+	default User selectExistedUser(long id) throws DBException {
 		try {
 			return selectUser(id);
 		}catch (NoSuchElement e) {
