+++ src/main/java/manager/dao/career/impl/WorkDAOImpl.java
@@ -40,27 +40,27 @@ import manager.util.TimeUtil;
 public class WorkDAOImpl implements WorkDAO {
 	private final SessionFactory hbFactory = getHibernateSessionFactory();
 	@Override
-	public int insertPlan(Plan plan) throws DBException {
+	public long insertPlan(Plan plan) throws DBException {
 		return insertEntity(plan, hbFactory);
 	}
 
 	@Override
-	public int insertWorkSheet(WorkSheet sheet) throws DBException {
+	public long insertWorkSheet(WorkSheet sheet) throws DBException {
 		return insertEntity(sheet, hbFactory);
 	}
 	
 	@Override
-	public int insertPlanDept(PlanDept dept) throws DBException {
+	public long insertPlanDept(PlanDept dept) throws DBException {
 		return insertEntity(dept, hbFactory);
 	}
 
 	@Override
-	public Plan selectPlan(int id) throws NoSuchElement, DBException {
+	public Plan selectPlan(long id) throws NoSuchElement, DBException {
 		return selectEntity(id,Plan.class, hbFactory);
 	}
 
 	@Override
-	public WorkSheet selectWorkSheet(int id) throws NoSuchElement, DBException {
+	public WorkSheet selectWorkSheet(long id) throws NoSuchElement, DBException {
 		return selectEntity(id, WorkSheet.class, hbFactory);
 	}
 
@@ -70,22 +70,22 @@ public class WorkDAOImpl implements WorkDAO {
 	}
 
 	@Override
-	public List<Plan> selectPlansByOwnerAndStates(int ownerId,List<PlanState> states) throws DBException {
+	public List<Plan> selectPlansByOwnerAndStates(long ownerId,List<PlanState> states) throws DBException {
 		return selectEntitiesByFieldAndManyField(Plan.class,SMDB.F_OWNER_ID,ownerId, SMDB.F_STATE, states,PlanState::getDbCode,hbFactory);
 	}
 
 	@Override
-	public void deleteExistedPlan(int planId) throws DBException {
+	public void deleteExistedPlan(long planId) throws DBException {
 		deleteEntity(Plan.class, planId, hbFactory);
 	}
 	
 	@Override
-	public long countPlansByOwnerAndState(int ownerId, PlanState state) throws DBException {
+	public long countPlansByOwnerAndState(long ownerId, PlanState state) throws DBException {
 		return countEntitiesByBiFields(Plan.class,SMDB.F_OWNER_ID,ownerId, SMDB.F_STATE,state.getDbCode(), hbFactory);
 	}
 
 	@Override
-	public long countWorkSheetByOwnerAndState(int ownerId, WorkSheetState state) throws DBException {
+	public long countWorkSheetByOwnerAndState(long ownerId, WorkSheetState state) throws DBException {
 		return countEntitiesByBiFields(WorkSheet.class,SMDB.F_OWNER_ID,ownerId, SMDB.F_STATE,state.getDbCode(), hbFactory);
 	}
 	
@@ -95,7 +95,7 @@ public class WorkDAOImpl implements WorkDAO {
 	}
 
 	@Override
-	public boolean includeUniqueWorkSheetByOwnerAndDate(int ownerId, Calendar date) throws DBException {
+	public boolean includeUniqueWorkSheetByOwnerAndDate(long ownerId, Calendar date) throws DBException {
 		return includeUniqueEntityByBiFields(WorkSheet.class,SMDB.F_OWNER_ID,ownerId,SMDB.F_DATE,date,hbFactory);
 	}
 
@@ -105,12 +105,12 @@ public class WorkDAOImpl implements WorkDAO {
 	}
 
 	@Override
-	public void deleteExistedWorkSheet(int wsId) throws DBException {
+	public void deleteExistedWorkSheet(long wsId) throws DBException {
 		deleteEntity(WorkSheet.class, wsId, hbFactory);
 	}
 
 	@Override
-	public List<WorkSheet> selectWorkSheetInfoRecentlyByOwner(int ownerId, int page,int limit) throws DBException {
+	public List<WorkSheet> selectWorkSheetInfoRecentlyByOwner(long ownerId, long page,long limit) throws DBException {
 		List<WorkSheet> rlt = new ArrayList<WorkSheet>();
 		Session session = null;
 		Transaction trans = null;
@@ -122,12 +122,12 @@ public class WorkDAOImpl implements WorkDAO {
 						SMDB.F_ID,SMDB.F_DATE,SMDB.F_STATE,
 						SMDB.T_WORK_SHEET, SMDB.F_OWNER_ID,SMDB.F_DATE,(page*limit),limit);
 				try (PreparedStatement ps = conn.prepareStatement(sql)) {
-					ps.setInt(1, ownerId);
+					ps.setLong(1, ownerId);
 					
 					ResultSet rs = ps.executeQuery();
 					while(rs.next()) {
 						WorkSheet ws =new WorkSheet();
-						ws.setId(rs.getInt(1));
+						ws.setId(rs.getLong(1));
 						Calendar date = Calendar.getInstance();
 						date.setTime(rs.getDate(2));
 						ws.setDate(date);
@@ -149,12 +149,12 @@ public class WorkDAOImpl implements WorkDAO {
 	}
 
 	@Override
-	public boolean includeWorkSheetByPlanId(int planId) throws DBException {
+	public boolean includeWorkSheetByPlanId(long planId) throws DBException {
 		return includeEntitiesByField(WorkSheet.class, SMDB.F_PLAN_ID, planId, hbFactory);
 	}
 
 	@Override
-	public PlanDept selectPlanDeptByOwner(int ownerId) throws NoSuchElement, DBException {
+	public PlanDept selectPlanDeptByOwner(long ownerId) throws NoSuchElement, DBException {
 		return selectUniqueEntityByField(PlanDept.class, SMDB.F_OWNER_ID, ownerId, hbFactory);
 	}
 
@@ -169,13 +169,13 @@ public class WorkDAOImpl implements WorkDAO {
 	}
 
 	@Override
-	public List<WorkSheet> selectWorkSheetByOwnerAndStates(int ownerId, List<WorkSheetState> states)
+	public List<WorkSheet> selectWorkSheetByOwnerAndStates(long ownerId, List<WorkSheetState> states)
 			throws DBException {
 		return selectEntitiesByFieldAndManyField(WorkSheet.class,SMDB.F_OWNER_ID,ownerId, SMDB.F_STATE, states,WorkSheetState::getDbCode,hbFactory);
 	}
 	
 	@Override
-	public List<String> selectNonNullPlanTagsByUser(int loginerId) throws DBException {
+	public List<String> selectNonNullPlanTagsByUser(long loginerId) throws DBException {
 		List<String> rlt = new ArrayList<>();
 		Session session = null;
 		Transaction trans = null;
@@ -187,7 +187,7 @@ public class WorkDAOImpl implements WorkDAO {
 						SMDB.F_TAGS,SMDB.T_PLAN,
 						SMDB.F_OWNER_ID,SMDB.F_TAGS,SMDB.F_TAGS);
 				try (PreparedStatement ps = conn.prepareStatement(sql)) {
-					ps.setInt(1, loginerId);
+					ps.setLong(1, loginerId);
 					ResultSet rs = ps.executeQuery();
 					while(rs.next()) {
 						rlt.add(rs.getString(1));
@@ -202,7 +202,7 @@ public class WorkDAOImpl implements WorkDAO {
 	}
 
 	@Override
-	public List<String> selectNonNullWorkSheetTagsByUser(int loginerId) throws DBException {
+	public List<String> selectNonNullWorkSheetTagsByUser(long loginerId) throws DBException {
 		List<String> rlt = new ArrayList<>();
 		Session session = null;
 		Transaction trans = null;
@@ -214,7 +214,7 @@ public class WorkDAOImpl implements WorkDAO {
 						SMDB.F_TAGS,SMDB.T_WORK_SHEET,
 						SMDB.F_OWNER_ID,SMDB.F_TAGS,SMDB.F_TAGS);
 				try (PreparedStatement ps = conn.prepareStatement(sql)) {
-					ps.setInt(1, loginerId);
+					ps.setLong(1, loginerId);
 					ResultSet rs = ps.executeQuery();
 					while(rs.next()) {
 						rlt.add(rs.getString(1));
@@ -230,7 +230,7 @@ public class WorkDAOImpl implements WorkDAO {
 	
 
 	@Override
-	public List<Plan> selectPlanInfosByIds(List<Integer> planIds) throws DBException {
+	public List<Plan> selectPlanInfosByIds(List<Long> planIds) throws DBException {
 		if(planIds.size() == 0) {
 			return new ArrayList<Plan>();
 		}
@@ -250,7 +250,7 @@ public class WorkDAOImpl implements WorkDAO {
 					ResultSet rs = ps.executeQuery();
 					while(rs.next()) {
 						Plan one =new Plan();
-						one.setId(rs.getInt(1));
+						one.setId(rs.getLong(1));
 						one.setName(rs.getString(2));
 						rlt.add(one);
 					}
@@ -264,12 +264,12 @@ public class WorkDAOImpl implements WorkDAO {
 	}
 
 	@Override
-	public long countWorkSheetByOwnerAndPlanId(int ownerId, int planId) throws DBException {
+	public long countWorkSheetByOwnerAndPlanId(long ownerId, long planId) throws DBException {
 		return countEntitiesByBiFields(WorkSheet.class,SMDB.F_OWNER_ID,ownerId, SMDB.F_PLAN_ID,planId, hbFactory);
 	}
 
 	@Override
-	public List<WorkSheet> selectWorkSheetsByOwnerAndDateScope(int ownerId, Calendar startDate, Calendar endDate) throws DBException {
+	public List<WorkSheet> selectWorkSheetsByOwnerAndDateScope(long ownerId, Calendar startDate, Calendar endDate) throws DBException {
 		return selectEntitiesByDateScopeAndField(WorkSheet.class, SMDB.F_DATE, startDate, endDate, SMDB.F_OWNER_ID, ownerId, hbFactory);
 	}
