+++ src/main/java/manager/entity/general/career/Memo.java
@@ -23,7 +23,7 @@ public class Memo extends SMGeneralEntity {
 	private String content;
 	
 	@Column
-	private Integer ownerId;
+	private Long ownerId;
 
 	@Column
 	private String note;
@@ -36,11 +36,11 @@ public class Memo extends SMGeneralEntity {
 		this.content = content;
 	}
 
-	public Integer getOwnerId() {
+	public Long getOwnerId() {
 		return ownerId;
 	}
 
-	public void setOwnerId(Integer ownerId) {
+	public void setOwnerId(Long ownerId) {
 		this.ownerId = ownerId;
 	}
